+++ includes/session.php
@@ -141,7 +141,7 @@ class Session {
         }
 
         // The operator doesn't exist locally. Check externally.
-        if(externalOperatorCheck($operatorId)) {
+        if(limsRequest(["operatorId" => $operatorId])) {
             // If the operator exists externally, create a clinician and log them in.
             $cainDB->query("INSERT INTO `users` (`operator_id`, `user_type`) VALUES (:operatorId, 1);", [':operatorId' => $operatorId]);